src/backend/requirements.txt
#### Requirements without Version Specified #### 

openai
flask
flask_cors
flask-apispec
flask-restful
gunicorn
python-dotenv